pyproject.toml
[tool.poetry]
name = "know_direction"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastkml = "^0.11"
lxml = "^4.6.3"
attrs = "^21.2.0"
networkx = "^2.5.1"
more-itertools = "^8.8.0"
black = "^21.6b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
